Organize all components into folders
Move App.js to src/
Update index.js to import App.js from /src
